Report average latency and throughput in xrt::runner w. profile (#8904)
* Report average latency and throughput in xrt::runner w. profile

Default report average latency and throughput when running a recipe
with a profile.

The runner implementation is fixed to handle iterations with minimum
wait.  For example, when the runlist consists of indvidually managed
xrt::run objects, then a iteration restarts all the xrt::run objects
waiting on runs in FIFO manner and restarting ASAP.

Latency is measured as expired time divided by the product of number of runs in the runlist and number of iterations.

Throughput is measured as the product of number of runs in runlist and
number of iterations divided by the expired time.

This is generic and optimal throughput and latency for current XRT
execution.  The recipe can be modifiled to include as many runs as
desired in order to keep the hardware busy.

There is no need to specify iteration wait in the profile json, in
fact if wait is specified, the granularity of wait is at the recipe
level, meaning that an recipe execution must complete before it be
iterated.

This commit also has changes to xrt_core:: timing which is changed to
use `unsigned long long` for timers.


* Fix fallout from xrt_core::time change


---------

+++ src/runtime_src/core/common/time.cpp
@@ -42,13 +42,13 @@ namespace xrt_core {
  * @return
  *   nanoseconds since first call
  */
-unsigned long
+unsigned long long
 time_ns()
 {
   static auto zero = std::chrono::high_resolution_clock::now();
   auto now = std::chrono::high_resolution_clock::now();
   auto integral_duration = std::chrono::duration_cast<std::chrono::nanoseconds>(now-zero).count();
-  return static_cast<unsigned long>(integral_duration);
+  return static_cast<unsigned long long>(integral_duration);
 }
 
 /**
